+++ packages/bigby-2d-editor/src/index.ts
@@ -1,0 +1,11 @@
+import EditorGrid2D from "./EditorGrid2D";
+import SelectedEntityGizmo2D from "./SelectedEntityGizmo2D";
+import ViewportController2D from "./ViewportController2D";
+import with2DEditor from "./with2DEditor";
+
+export {
+  EditorGrid2D,
+  SelectedEntityGizmo2D,
+  ViewportController2D,
+  with2DEditor,
+};