+++ packages/bigby-2d/src/TilingSprite2D.ts
@@ -1,8 +1,8 @@
 import { inspect } from "@bigby/core";
 import * as PIXI from "pixi.js";
-import { Sprite2D } from "./Sprite2D";
+import Sprite2D from "./Sprite2D";
 
-export class TilingSprite2D extends Sprite2D<{
+export default class TilingSprite2D extends Sprite2D<{
   bounds: { width: number; height: number };
 }> {
   static displayName = "TilingSprite2D";