+++ packages/bigby-3d-editor/src/index.ts
@@ -1,0 +1,3 @@
+import with3DEditor from "./with3DEditor";
+
+export { with3DEditor };