+++ packages/bigby-3d/src/index.ts
@@ -1,0 +1,3 @@
+import Game3D from "./Game3D";
+
+export { Game3D };