+++ packages/bigby-behaviors/src/index.ts
@@ -1,4 +1,5 @@
 import KeyboardInput from "./KeyboardInput";
+import Ticker from "./Ticker";
 import Timer from "./Timer";
 
-export { Timer, KeyboardInput };
+export { Ticker, Timer, KeyboardInput };