+++ packages/bigby-editor/rollup.config.js
@@ -5,10 +5,10 @@ import pkg from "./package.json";
 export default {
   input: "src/index.ts",
   output: [
-    {
-      file: pkg.module,
-      format: "es", // the preferred format
-    },
+    // {
+    //   file: pkg.module,
+    //   format: "es", // the preferred format
+    // },
     {
       file: pkg.main,
       format: "cjs",